+++ tests/test_simple_module.py
@@ -95,12 +95,12 @@ def test_basic_module(project: Project) -> None:
         name="tests",
         path=[module.name],
         entity=entity,
-        arity=(0, None),
+        cardinality=(0, None),
         peer=EntityRelation(
             name="",
             path=[module.name],
             entity=entity,
-            arity=(0, 0),
+            cardinality=(0, 0),
         ),
     )