+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v 1.0.0
 
+ - Renaming `arity` to `cardinality` in relations
  - Add support for relations in index (breaking change) (#101)
  - Extend list of inmanta reserved keywords